# iscc_search/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ISCC-VDB API
  description: A Scalable Nearest Neighbor Search Multi-Index for the International Standard Content Code (ISCC)
  version: 0.1.0
servers:
  - url: http://localhost:8000
    description: Development server

paths:
  /indexes:
    get:
      summary: List all indexes
      description: Returns a list of all available ISCC indexes with their metadata
      operationId: listIndexes
      tags:
        - indexes
      responses:
        '200':
          description: Successful response with list of indexes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IsccIndex'
              example:
                - name: "default"
                  assets: 150000
                  size: 42
    post:
      summary: Create a new index
      description: Creates a new ISCC index with the specified name. The index is initialized empty with 0 assets.
      operationId: createIndex
      tags:
        - indexes
      requestBody:
        required: true
        description: Index configuration with unique name
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IsccIndex'
            example:
              name: "myindex"
      responses:
        '201':
          description: Index successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IsccIndex'
              example:
                name: "myindex"
                assets: 0
                size: 0
        '400':
          description: Invalid request (e.g., invalid name format)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
              example:
                detail: "Invalid index name format"
        '409':
          description: Index with this name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
              example:
                detail: "Index 'myindex' already exists"
  /indexes/{name}:
    get:
      summary: Get a specific index
      description: Returns metadata for a specific ISCC index by name
      operationId: getIndex
      tags:
        - indexes
      parameters:
        - name: name
          in: path
          required: true
          description: Name of the index to retrieve
          schema:
            type: string
            pattern: "^[a-z][a-z0-9]*$"
            minLength: 1
            maxLength: 32
            example: "default"
      responses:
        '200':
          description: Successful response with index metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IsccIndex'
              example:
                name: "default"
                assets: 150000
                size: 42
        '400':
          description: Invalid index name format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
              example:
                detail: "Invalid index name format"
        '404':
          description: Index not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
              example:
                detail: "Index 'myindex' not found"
    delete:
      summary: Delete an index
      description: Permanently deletes an ISCC index and all its data
      operationId: deleteIndex
      tags:
        - indexes
      parameters:
        - name: name
          in: path
          required: true
          description: Name of the index to delete
          schema:
            type: string
            pattern: "^[a-z][a-z0-9]*$"
            minLength: 1
            maxLength: 32
            example: "myindex"
      responses:
        '204':
          description: Index successfully deleted (no content returned)
        '400':
          description: Invalid index name format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
              example:
                detail: "Invalid index name format"
        '404':
          description: Index not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
              example:
                detail: "Index 'myindex' not found"
  /indexes/{name}/assets:
    post:
      summary: Add assets to an index
      description: |
        Adds multiple ISCC assets to the specified index. Assets are processed atomically
        (all-or-nothing). Each asset must include an iscc_id field (client-generated).
        Returns status for each asset indicating whether it was newly created or already existed.
      operationId: addAssets
      tags:
        - assets
      parameters:
        - name: name
          in: path
          required: true
          description: Name of the index to add assets to
          schema:
            type: string
            pattern: "^[a-z][a-z0-9]*$"
            minLength: 1
            maxLength: 32
            example: "default"
      requestBody:
        required: true
        description: Array of ISCC assets to add (either iscc_code or units required)
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/IsccAsset'
            example:
              - iscc_code: "ISCC:KECYCMZIOY36XXGZ7S6QJQ2AEEXPOVEHZYPK6GMSFLU3WF54UPZMTPY"
                units:
                  - "ISCC:AADYCMZIOY36XXGZ5B5BME7EIPPXRFKYQZ7VXKI7V55AEQQE67A33BY"
                  - "ISCC:EED7ZPIEYNACCLXXZSS2LIM6JVXDYGCG2QSMC7DCPER4MYJPJATIM4Y"
              - units:
                  - "ISCC:GADVJB6OD2XRTERK3G4KOFHGDJCBYEXAZCZLVVSA7MMURDPMJ5S5JWI"
                  - "ISCC:IAD6TOYXXSR7FSN7TRRRNOKQ6JCFK3ZF4KRFVEQRQ4M4PDPUR5H7GHQ"
      responses:
        '201':
          description: Assets successfully added to index
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IsccAddResult'
              example:
                - iscc_id: "ISCC:MAIGIIFJRDGEQQAA"
                  status: "created"
                - iscc_id: "ISCC:MAIGXXFZRDGEQQBB"
                  status: "created"
        '400':
          description: Invalid request (malformed ISCC, validation error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
              example:
                detail: "Invalid ISCC format in asset 1"
        '404':
          description: Index not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
              example:
                detail: "Index 'myindex' not found"
  /indexes/{name}/assets/{iscc_id}:
    get:
      summary: Get a specific asset by ISCC-ID
      description: |
        Retrieves a single ISCC asset from the index by its ISCC-ID. This is useful for
        fetching full asset details after a search operation, which returns only ISCC-IDs.
      operationId: getAsset
      tags:
        - assets
      parameters:
        - name: name
          in: path
          required: true
          description: Name of the index containing the asset
          schema:
            type: string
            pattern: "^[a-z][a-z0-9]*$"
            minLength: 1
            maxLength: 32
            example: "default"
        - name: iscc_id
          in: path
          required: true
          description: ISCC-ID of the asset to retrieve
          schema:
            type: string
            pattern: "^ISCC:[A-Z2-7]{16}$"
            example: "ISCC:MAIGIIFJRDGEQQAA"
      responses:
        '200':
          description: Successful response with asset details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IsccAsset'
              example:
                iscc_id: "ISCC:MAIGIIFJRDGEQQAA"
                iscc_code: "ISCC:KECYCMZIOY36XXGZ7S6QJQ2AEEXPOVEHZYPK6GMSFLU3WF54UPZMTPY"
                units:
                  - "ISCC:AADYCMZIOY36XXGZ5B5BME7EIPPXRFKYQZ7VXKI7V55AEQQE67A33BY"
                  - "ISCC:EED7ZPIEYNACCLXXZSS2LIM6JVXDYGCG2QSMC7DCPER4MYJPJATIM4Y"
                metadata:
                  title: "Example Document"
                  creator: "John Doe"
        '400':
          description: Invalid ISCC-ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
              example:
                detail: "Invalid ISCC-ID format"
        '404':
          description: Index or asset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
              examples:
                indexNotFound:
                  value:
                    detail: "Index 'myindex' not found"
                assetNotFound:
                  value:
                    detail: "Asset 'ISCC:MAIGIIFJRDGEQQAA' not found in index 'myindex'"
  /indexes/{name}/search:
    post:
      summary: Search for similar ISCCs (POST)
      description: |
        Searches for similar ISCC assets in the index using the provided query asset.
        Supports complex queries with multiple units and custom ISCC-CODEs.
      operationId: searchPost
      tags:
        - search
      parameters:
        - name: name
          in: path
          required: true
          description: Name of the index to search
          schema:
            type: string
            pattern: "^[a-z][a-z0-9]*$"
            minLength: 1
            maxLength: 32
            example: "default"
      requestBody:
        required: true
        description: ISCC asset to use as search query
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IsccAsset'
            example:
              iscc_code: "ISCC:KECYCMZIOY36XXGZ7S6QJQ2AEEXPOVEHZYPK6GMSFLU3WF54UPZMTPY"
              units:
                - "ISCC:AADYCMZIOY36XXGZ5B5BME7EIPPXRFKYQZ7VXKI7V55AEQQE67A33BY"
                - "ISCC:EED7ZPIEYNACCLXXZSS2LIM6JVXDYGCG2QSMC7DCPER4MYJPJATIM4Y"
      responses:
        '200':
          description: Search results with matched assets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IsccSearchResult'
              example:
                query:
                  iscc_code: "ISCC:KECYCMZIOY36XXGZ7S6QJQ2AEEXPOVEHZYPK6GMSFLU3WF54UPZMTPY"
                  units:
                    - "ISCC:AADYCMZIOY36XXGZ5B5BME7EIPPXRFKYQZ7VXKI7V55AEQQE67A33BY"
                    - "ISCC:EED7ZPIEYNACCLXXZSS2LIM6JVXDYGCG2QSMC7DCPER4MYJPJATIM4Y"
                metric: "bitlength"
                matches:
                  - iscc_id: "ISCC:MAIGIIFJRDGEQQAA"
                    score: 448
                    matches:
                      CONTENT_TEXT_V0: 256
                      DATA_NONE_V0: 128
                      INSTANCE_NONE_V0: 64
        '400':
          description: Invalid request (malformed ISCC)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
              example:
                detail: "Invalid ISCC format"
        '404':
          description: Index not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
              example:
                detail: "Index 'myindex' not found"
    get:
      summary: Search for similar ISCCs (GET)
      description: |
        Convenience endpoint for searching by ISCC-CODE via query parameter.
        The ISCC-CODE is decomposed into units for searching.
      operationId: searchGet
      tags:
        - search
      parameters:
        - name: name
          in: path
          required: true
          description: Name of the index to search
          schema:
            type: string
            pattern: "^[a-z][a-z0-9]*$"
            minLength: 1
            maxLength: 32
            example: "default"
        - name: iscc_code
          in: query
          required: true
          description: ISCC-CODE to search for (will be decomposed into units)
          schema:
            type: string
            pattern: "^ISCC:[A-Z2-7]{16,}$"
          example: "ISCC:KECYCMZIOY36XXGZ7S6QJQ2AEEXPOVEHZYPK6GMSFLU3WF54UPZMTPY"
      responses:
        '200':
          description: Search results with matched assets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IsccSearchResult'
              example:
                query:
                  iscc_code: "ISCC:KECYCMZIOY36XXGZ7S6QJQ2AEEXPOVEHZYPK6GMSFLU3WF54UPZMTPY"
                  units:
                    - "ISCC:AADYCMZIOY36XXGZ5B5BME7EIPPXRFKYQZ7VXKI7V55AEQQE67A33BY"
                    - "ISCC:EED7ZPIEYNACCLXXZSS2LIM6JVXDYGCG2QSMC7DCPER4MYJPJATIM4Y"
                metric: "bitlength"
                matches:
                  - iscc_id: "ISCC:MAIGIIFJRDGEQQAA"
                    score: 448
                    matches:
                      CONTENT_TEXT_V0: 256
                      DATA_NONE_V0: 128
                      INSTANCE_NONE_V0: 64
        '400':
          description: Invalid ISCC-CODE format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
              example:
                detail: "Invalid ISCC-CODE format"
        '404':
          description: Index not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpError'
              example:
                detail: "Index 'myindex' not found"

components:
  schemas:
    HttpError:
      $ref: './HttpError.yaml'
    IsccIndex:
      $ref: './IsccIndex.yaml'
    IsccAsset:
      $ref: './IsccAsset.yaml'
    IsccAddResult:
      $ref: './IsccAddResult.yaml'
    IsccMatch:
      $ref: './IsccMatch.yaml'
    IsccSearchResult:
      $ref: './IsccSearchResult.yaml'